Unified `u32` type for all screen size manipulations in miniquad

+++ src/lib.rs
@@ -115,9 +115,9 @@ pub mod window {
 	/// The current framebuffer size in pixels.
 	/// If [`Conf::high_dpi``](crate::conf::Conf::high_dpi) was set to false, canvas|window size == framebuffer size.
 	/// If set to true, the framebuffer is scaled by the dpi.
-	pub fn screen_size() -> (f32, f32) {
+	pub fn screen_size() -> (u32, u32) {
 		let d = native_display().lock().unwrap();
-		(d.screen_width as f32, d.screen_height as f32)
+		(d.screen_width as u32, d.screen_height as u32)
 	}
 
 	/// The dpi scaling factor (window pixels to framebuffer pixels)
